json parser went to mir_core
git-svn-id: http://svn.miranda-ng.org/main/trunk@5471 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ src/mir_core/json/JSONDefs.h
@@ -65,7 +65,7 @@
 #ifdef JSON_INDEX_TYPE
 	typedef JSON_INDEX_TYPE json_index_t;
 #else
-	typedef unsigned int json_index_t;
+	typedef size_t json_index_t;
 #endif
 
 typedef void (*json_mutex_callback_t)(void *);
@@ -76,8 +76,8 @@ typedef void (*json_free_t)(void *);
 #else
 	#define JSONNODE void  //so that JSONNODE* is void*
 	typedef JSONNODE** JSONNODE_ITERATOR;
-	typedef void * (*json_malloc_t)(unsigned long);
-	typedef void * (*json_realloc_t)(void *, unsigned long);
+	typedef void * (*json_malloc_t)(size_t);
+	typedef void * (*json_realloc_t)(void *, size_t);
 #endif
 
 #endif //JSONDEFS_H
